pyproject.toml
[project]
name = "python-teaching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.25.1",
    "beautifulsoup4>=4.14.2",
    "click>=8.3.0",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "jupyter-book>=1.0.4.post1",
    "matplotlib>=3.10.7",
    "mypy>=1.18.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pypdf>=6.1.2",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "ruff>=0.14.1",
    "seaborn>=0.13.2",
    "typer>=0.19.2",
]
